requirements.txt
keras
numpy
python-chess
scikit-learn
tensorflow
tqdm
